fix(visibleWidgets): track visibleWidget changes in React-compatible way
When hiding and showing widgets on the main app screen, we are using Javascript `Set` to track what widgets are visible. Default React implementation of dependency analysis for complex types like `Set` or `Array` compares just the reference to see if `Set` or `Array` changed. Track changes to our set of visible widgets and when we detect a change, force create new Javascript `Set` reference to make React correctly re-render the page.