feat: Introduce OffscreenPageSplitter
When we have a long list of children that we need to split into pages,
we need to lay them out with aplied CSS and precisely measure them
to compute the pages. This must respect the fact that children that fit
in one page need not to display page selector, but in all other cases
page selector is visible thus reducing the available space.

Children are either laid out vertically one below each other,
or horizontally side by side.

Available space represents how much space is available in the layout dimension.

OffscerenPageSplitter takes the children and lays them out in a hidden flexbox,
in the respective direction, including the hidden page selector. It then measures
the children sizes and page selector size and invokes a user defined callback.

This logic re-runs automatically as the CSS is applied to the children,
list of children changes, or viewport size changes.

Previously we simply rendered the children directly and observed their size
and scroll size of their container in order to determine whether we need paging.

But this approach was unstable for multiple reasons. Rendering all children onscreen
and then hiding them when paging was needed caused flicker. Resizeing viewport caused flicker.
Showing paging indicator when paging started caused re-rendering that caused re-computation of
the paging information causing flickering in certain edge cases.

OffscreenPageSplitter always renders offscreen first, and only when the exact paging information
is computed it passed it up via callback and performs exact rendering of the pages.