feat: Refactor Paginator to use OffscreenPageSplitter
Paginator used to work in a way where it would draw its children
in a flexbox laying them out vertically or horizontally. The children
were wrapped in SizeChangeObserver to recompute paging when their size
changes because of incremental CSS changes, and rendered inside container
that used ScrollSizeObserver listening for whether the paging is needed
(container would overflow and need to scroll).

This caused lot of flicker and instability, as elements were dreawn,
and then hidden, as CSS can get applied slowly on older devices,
and as optionally showing page selector would trigger the
scroll size change obsever and trigger the re-computing again causing
flicker in the best case, and un-solvable layout in worst case.

New implementation uses OffscreenPageSpliter where it first renders
all children offscreen and measures them, and based on the real sizes
computes stable paging information that is later used to draw
exact pages. This resolves the flicker and makes paging smooth to recompute
should CSS or viewport size change.

Closes #421